Don't use `as` for casting pointers (only for `*const T -> *mut T`)
Easy to accidentally convert immutable pointers to a mutable ones, as is done in `declare` (now fixed).

+++ objc2-foundation/src/enumerator.rs
@@ -172,25 +172,25 @@ mod tests {
 
     #[test]
     fn test_enumerator() {
-        let vec = (0u32..4).map(NSValue::new).collect();
+        let vec = (0usize..4).map(NSValue::new).collect();
         let array = NSArray::from_vec(vec);
 
         let enumerator = array.iter();
         assert_eq!(enumerator.count(), 4);
 
         let enumerator = array.iter();
-        assert!(enumerator.enumerate().all(|(i, obj)| obj.get() == i as u32));
+        assert!(enumerator.enumerate().all(|(i, obj)| obj.get() == i));
     }
 
     #[test]
     fn test_fast_enumerator() {
-        let vec = (0u32..4).map(NSValue::new).collect();
+        let vec = (0usize..4).map(NSValue::new).collect();
         let array = NSArray::from_vec(vec);
 
         let enumerator = array.iter_fast();
         assert_eq!(enumerator.count(), 4);
 
         let enumerator = array.iter_fast();
-        assert!(enumerator.enumerate().all(|(i, obj)| obj.get() == i as u32));
+        assert!(enumerator.enumerate().all(|(i, obj)| obj.get() == i));
     }
 }
